+++ tests/testthat/test-instants.R
@@ -4,18 +4,18 @@ test_that("is.instant/is.timepoint works as expected",{
   expect_that(is.instant(234), is_false())
   expect_that(is.instant(as.POSIXct("2008-08-03 13:01:59", tz = "UTC")),
     is_true())
-  expect_that(is.instant(as.POSIXlt("2008-08-03 13:01:59", tz = "UTC")), 
+  expect_that(is.instant(as.POSIXlt("2008-08-03 13:01:59", tz = "UTC")),
     is_true())
   expect_that(is.instant(Sys.Date()), is_true())
   expect_that(is.instant(minutes(1)), is_false())
   expect_that(is.timespan(interval(
-    as.POSIXct("2008-08-03 13:01:59", tz = "UTC"), 
+    as.POSIXct("2008-08-03 13:01:59", tz = "UTC"),
     as.POSIXct("2009-08-03 13:01:59", tz = "UTC") )), is_true())
 })
 
 test_that("is.instant/is.timepoint handle vectors",{
   expect_that(is.instant(minutes(1:2)), is_false())
-  expect_that(is.instant(as.POSIXct(c("2008-08-03 13:01:59", 
+  expect_that(is.instant(as.POSIXct(c("2008-08-03 13:01:59",
   	"2008-08-03 13:01:59"), tz = "UTC")), is_true())
 })
 
@@ -32,7 +32,7 @@ test_that("today() works correctly",{
 test_that("make_datetime returns same values as ISOdatetime", {
 
   set.seed(1000)
-  
+
   N <- 1e4
   y <- as.integer(runif(N, 1800, 2200))
   m <- as.integer(runif(N, 1, 12))
@@ -50,6 +50,9 @@ test_that("make_datetime returns same values as ISOdatetime", {
   out1 <- ISOdatetime(y, m, d, H, M, S, tz = "UTC")
   out2 <- make_datetime(y, m, d, H, M, S)
   expect_equal(out1, out2)
+
+  out3 <- make_date(y, m, d)
+  expect_equal(as.Date(out1), out3)
 })
 
 test_that("make_datetime replicates as expected", {
@@ -58,6 +61,6 @@ test_that("make_datetime replicates as expected", {
 })
 
 test_that("make_datetime propagates NAs as expected", {
-  expect_equal(make_datetime(year = 1999, month = c(11, NA), day = 22, sec = c(10, 11, NA)), 
+  expect_equal(make_datetime(year = 1999, month = c(11, NA), day = 22, sec = c(10, 11, NA)),
                as.POSIXct(c("1999-11-22 00:00:10 UTC", NA, NA), tz = "UTC"))
 })